refactor: share structs in shared crate
