+++ crates/shared/Cargo.toml
@@ -1,14 +1,9 @@
 [package]
-name          = "inspect_core"
+name          = "shared"
 license       = "MIT"
 repository    = "https://github.com/pmqueiroz/import-holmes.git"
 version       = "0.1.0"
 
 [dependencies]
-shared = { path = "../shared" }
 serde = { version = "1.0.213", features = ["derive"] }
 serde_json = "1.0.132"
-swc_common = "0.29.37"
-swc_ecma_ast = "0.100.1"
-swc_ecma_parser = "0.130.3"
-swc_ecma_visit = "0.86.1"
